third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-video-resize.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

function generateTest()
{
    return Promise.all([
        createImageBitmap(video, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "high"}),
        createImageBitmap(video, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "medium"}),
        createImageBitmap(video, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "low"}),
        createImageBitmap(video, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "pixelated"}),
        createImageBitmap(video, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "high"}),
        createImageBitmap(video, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "medium"}),
        createImageBitmap(video, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "low"}),
        createImageBitmap(video, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "pixelated"}),
    ]).then(t.step_func_done(([noCropHigh, noCropMedium, noCropLow, noCropPixelated, cropHigh, cropMedium, cropLow, cropPixelated]) => {
    }), t.step_func_done(function() {
        assert_true(false, 'Promise rejected');
    }));
}

var t = async_test('createImageBitmap(HTMLVideoElement) with resize option');

// HTMLVideoElement
var video = document.createElement("video");
video.oncanplaythrough = t.step_func(function() {
    return generateTest();
});
video.src = "../../compositing/resources/video.ogv";
</script>
